Fix nft summary
