+++ src/components/v2v3/V2V3Project/V2V3ProjectSettings/pages/ProjectUpgradesPage/versions/upgradeForms/V3/UpgradeFundingCycle/LaunchFundingCycleForm.tsx
@@ -16,7 +16,7 @@ import { useDispatch } from 'react-redux'
 import { editingV2ProjectActions } from 'redux/slices/editingV2Project'
 import { formatDate } from 'utils/format/formatDate'
 import { useEditingFundingCycleConfig } from '../../../../../ReconfigureFundingCycleSettingsPage/hooks/editingFundingCycleConfig'
-import ReconfigurePreview from '../../../../../ReconfigureFundingCycleSettingsPage/ReconfigurePreview'
+import ReconfigurePreview from '../../../../../ReconfigureFundingCycleSettingsPage/ReconfigurePreview/ReconfigurePreview'
 import { SetNftOperatorPermissionsButton } from '../../../../../ReconfigureFundingCycleSettingsPage/SetNftOperatorPermissionsButton'
 import { useLaunchFundingCycles } from './hooks/LaunchFundingCycles'
