+++ src/components/v2v3/V2V3Project/banners/RelaunchFundingCycleBanner/RelaunchFundingCycleModal.tsx
@@ -27,7 +27,7 @@ import {
 } from 'models/v2v3/fundingCycle'
 import { useContext, useEffect, useMemo, useState } from 'react'
 import { reloadWindow } from 'utils/windowUtils'
-import ReconfigurePreview from '../../V2V3ProjectSettings/pages/ReconfigureFundingCycleSettingsPage/ReconfigurePreview'
+import ReconfigurePreview from '../../V2V3ProjectSettings/pages/ReconfigureFundingCycleSettingsPage/ReconfigurePreview/ReconfigurePreview'
 
 export function RelaunchFundingCycleModal(props: ModalProps) {
   const { projectId } = useContext(ProjectMetadataContext)
